+++ lib/src/models/models.dart
@@ -1,1 +1,2 @@
 part 'user.dart';
+part 'message.dart';
